requirements.txt
streamlit
google-generativeai
langchain
langchain-community
chromadb
PyMuPDF
pdfplumber
pillow
nougat-ocr
tiktoken
python-dotenv
